pyproject.toml
[project]
name = "email-qbit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]
